# chore: Reduce the amount of bytes allocated for Filter