Reuse parsed data structure for subsequent queries. See `B/op`


# Before

```
cpu: Intel(R) Core(TM) i7-6770HQ CPU @ 2.60GHz
BenchmarkRBACFilter/NoRoles-8             701658              1769 ns/op             138 B/op          3 allocs/op
BenchmarkRBACFilter/Admin-8              5158192               627.9 ns/op           291 B/op          0 allocs/op
BenchmarkRBACFilter/OrgAdmin-8             21584             53444 ns/op           16531 B/op        318 allocs/op
BenchmarkRBACFilter/OrgMember-8            22766             49841 ns/op           14289 B/op        274 allocs/op
BenchmarkRBACFilter/ManyRoles-8            29150             37300 ns/op           13166 B/op        254 allocs/op
BenchmarkRBACFilter/AdminWithScope-8      741349              1623 ns/op             138 B/op          3 allocs/op
```

# After

```
cpu: Intel(R) Core(TM) i7-6770HQ CPU @ 2.60GHz
BenchmarkRBACFilter/NoRoles-8             750855              1623 ns/op             138 B/op          3 allocs/op
BenchmarkRBACFilter/Admin-8              5281387               584.8 ns/op           284 B/op          0 allocs/op
BenchmarkRBACFilter/OrgAdmin-8             32541             34499 ns/op           10972 B/op        208 allocs/op
BenchmarkRBACFilter/OrgMember-8            27338             41346 ns/op           10602 B/op        201 allocs/op
BenchmarkRBACFilter/ManyRoles-8            50398             22765 ns/op            8819 B/op        168 allocs/op
BenchmarkRBACFilter/AdminWithScope-8      754546              1525 ns/op             138 B/op          3 allocs/op
```